[CXF-6634] Moving DocumentationProvider to a dedicated package
