+++ rt/rs/description/src/main/java/org/apache/cxf/jaxrs/model/doc/DocumentationProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.cxf.jaxrs.model.wadl;
+package org.apache.cxf.jaxrs.model.doc;
 
 import org.apache.cxf.jaxrs.model.ClassResourceInfo;
 import org.apache.cxf.jaxrs.model.OperationResourceInfo;
